保存病历

+++ src/main/java/com/littledoctor/clinicassistant/module/record/dao/MedicalRecordRepository.java
@@ -9,6 +9,6 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
  * @Date: 2019-08-24 10:27
  * @Description:
  */
-public interface MedicalRecordRepository extends JpaRepository<MedicalRecord, Integer>, JpaSpecificationExecutor<MedicalRecord> {
+public interface MedicalRecordRepository extends JpaRepository<MedicalRecord, Long>, JpaSpecificationExecutor<MedicalRecord> {
 
 }
